Move side panels out of edit views
